last

+++ src/ApiGateways/Shopping.Aggregator/Models/BasketModel.cs
@@ -9,6 +9,6 @@ namespace Shopping.Aggregator.Models
     {
         public string UserName { get; set; }
         public List<BasketItemExtendedModel> Items { get; set; } = new List<BasketItemExtendedModel>();
-        public decimal TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
     }
 }
